Created new server.js file that creates the server for the App
If applied, this commit will have created a new server.js file that creates the server for the App.
In this file was imported three different Node.js modules, the 'http', 'fs' and 'url'.
Through the createServer() function from 'http' was created the server that request the user's Url. After this Url request is parsed ('q' variable) and responsed by the 'fs' that using the appendFile() function send user's information, such as the addr (Url accessed) and the Timestamp to the log.txt file.
Was created a condition that allows to open the documentation.html if the 'documentation' word is included in the pathname(main), if not it opens the 'index.html'.
Another function from 'fs' was used, readFile() function that allows to send back the information requested by the user.