Updated index.js with new Express syntax in the context of the old file server.js written in pure Node.js..
If applied, this commit will have updated the index.js file with a new Express syntax adapting the old file server.js written in pure Node.js.
Was created a variable that contains the top 10 movies. This json file will be opened in a new /movies file.
Now it's in use 'morgan' package that is responsible as a middleware function to the Logging.
The file 'public' is now in use through the 'express.static'. This file will contain all the static files, and with morgan they will be send back automatically.
Finally was added a handle error middleware function that informs the terminal when unanticipated errors happen.