Updated package.json with dependencie: uuid, devDependencies: nodemon and updated the Scripts with a new dev.
If applied, this commit will have updated the package.json with a new uuid dependencies that allows, automatically to create a new ID for a new user.
As well, a new nodemon devDependencies that refresh automatically the port in the terminal when saving the file.
Finally, the nodemon was added to Scripts as a dev to start when using the "npm run dev" instead of the pure Node.