Updated index.js with new responses, modules, package importation, endpoints adjustments and added a Data Base..
If applied, this commit will have import new models.js and mongoose defining a Business Logic that clarifies, organizes and translates the data from MongoDB to the server (and afterwords to the client).
All the responses from every endpoint was changed according the new Data Base and Models.js use.
Was added 2 new endpoints for a new collection named 'Actors'
The listened port number was changed to 8080.