Created a passport.js file that contains the LocalStrategy and JWTStrategy for the Authentication and Authorization.
If applied, this commit will have created a new passport.js file that contains code to enable certain means of authentication and authorization.
For the Strategies we need to import and use the tool, Passport.
As well was imported the Models.js file to create the strategies directly in the Users collection that contains Users data.