Updated with 'Brcrypt' creating a new hashPassword and validatePassword.
If applied, this commit will have updated the models.js importing a bcrypt module that encrypts the password. This is called Hashing.
Likewise, was added two new statements, one that generates the hashPassword (then used in index.js POST(creating a new user))and another that validatePassword inserted(used in passport.js in if condition).