Updated index.js file importing 'cors' and 'express-validator'. Afterwords updated the POST - /users/ and PUT - /user/:Username with new validator.
If applied, this commit will have imported the two packages 'cors' and 'express-validator'.
With Cors we will defined the allowedOrigins, where is  designated who can access the API.
With express-validator we will be able to add in POST - /users/ and PUT - /user/:Username endpoints new 'checks' that define new rules in each input field for creating and updating an User, creating, like this more security and preventing from evils attacks to the Server-Side.